Added Note and Planner Functionality

+++ app/src/main/java/com/TechInfinityStudios/smartplanner/Note.java
@@ -1,20 +1,19 @@
 package com.TechInfinityStudios.smartplanner;
 
 public class Note {
-    private int id;
+    private long id;
     private String title;
     private String text;
 
-    // Constructors, getters, and setters
 
-    public Note( int id, String title, String text) {
+    public Note(long id, String title, String text) {
         this.id = id;
         this.title = title;
         this.text = text;
 
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
     public String getTitle() {
